4条边调整为条件边
